requirements.txt
pandas>=2.2.0
numpy>=1.26.0
xgboost>=2.0.0
scikit-learn>=1.4.0
scipy>=1.12.0
requests>=2.31.0
python-dotenv>=1.0.0
kagglehub>=0.2.0